Everything works

+++ iterator_test.go
@@ -38,7 +38,7 @@ type tableMock struct {
 
 func (tm *tableMock) Smallest() []byte             { return tm.left }
 func (tm *tableMock) Biggest() []byte              { return tm.right }
-func (tm *tableMock) DoesNotHave(hash uint64) bool { return false }
+func (tm *tableMock) DoesNotHave(hash uint32) bool { return false }
 
 func TestPickTables(t *testing.T) {
 	opt := DefaultIteratorOptions
